requirements/production.txt
-r base.txt
django-celery==3.1.1
gunicorn==18.0
redis==2.9.1
django-redis-cache==0.10.2
pyelasticsearch==0.6.1
elasticutils==0.8.2
paste
